Added new classes and updated styles
The buttons in index.html have new classes: button-1 and button-2.
The style.css file was updated to add styling for the buttons.
The first button has a black background and white text.
The second button has a yellow background.
They both have a font-family of Helvetica, Times New Roman, and sans-serif as a fallback.
They also have a font-size of 28px.